Confirm that you will release any merged apps
We often end up in a situation where lots of gem version bumps have been
merged, but they haven't been released. Sometimes, this is because this
tool has been used to quickly merge a version bump across lots of
applications, where perhaps only some of the apps are actually affected.
The affected applications usually get released, but others end up with
changes on the main branch that aren't deployed to production.

This is very risky. When there are unexpected incompatibilities with a
version bump, we won't discover them close to the time of the merge when
someone has enough context to work out what's wrong. We'll discover them
weeks or months later, when some poor soul decides to release the
application to get a different, unrelated change out.

To try to discourage people from using this tool in inappropriate
situations, I've added an annoying confirmation message to the `merge`
and `merge_only` tasks. The user must type "I promise to release
everything I merge" before the tool will do the merges now.

I'm not joking about this. I genuinely think this will help nudge
people's incentives in the right direction. Some of my wording might be
a bit aggressive though - happy to work on that if people think it's
needed.